requirements.txt
# --- CORE SYSTEM ---
requests
numpy
colorama
pyaudio
python-dotenv  # For safely loading API keys

# --- LLM / REASONING ---
xai-sdk
tavily-python
langchain-community
sentence-transformers

#RAG
faiss-cpu

# --- ENVIRONMENT AWARENESS ---
watchdog

# --- BACKEND API (The Nervous System) ---
fastapi
uvicorn
python-multipart
websockets

# --- EARS (Hearing) ---
faster-whisper

# --- MOUTH (Speech) ---
# Critical: ONNX Runtime GPU is required for fast neural speech on RTX 3050
kokoro-onnx
sounddevice
onnxruntime-gpu

# --- EYES (Vision) ---
transformers
pillow
einops
accelerate
torch
torchvision
torchaudio